pyproject.toml
[project]
name = "tao-fomc"
version = "0.1.0"
description = "FOMC analysis tools"
authors = [
    {name = "Entropy Foundation"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "aptos-sdk (>=0.11.0,<0.12.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[project.scripts]
fomc-submit = "aptos_rate_submitter:cli"
fomc-verify = "verify_tx:cli"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
# Disable packaging the current project (no packages in repo).
package-mode = false
